Face “Threat Level” Recognition to Computer Generated Art
The combination of facial recognition technology and core machine learning through the use of apple’s own technology (Core Image) gives the ability to recognise facial features to determine if someone is “threatening”. The threatening case is done by obtaining a variety of outputs such as if both eyes are showing or if the subject is smiling. The program then will use each of these factors and calculate what the image is deemed to be, such as if each case was “false” then the output would be of a high threatening level. The creative process however is when the program determines what variable is outputted and can generate art through a Plist of stored pieces of art. Demining the art however is a rigorous process as there is quite a lot of research to determine what is “threatening” and how art could be deemed to be “threatening”.